Enum ethers_core::abi::token::Token

pub enum Token {
    Address(H160),
    FixedBytes(Vec<u8, Global>),
    Bytes(Vec<u8, Global>),
    Int(U256),
    Uint(U256),
    Bool(bool),
    String(String),
    FixedArray(Vec<Token, Global>),
    Array(Vec<Token, Global>),
    Tuple(Vec<Token, Global>),
}

Ethereum ABI params.

Variants

Address(H160)

Address.

solidity name: address Encoded to left padded [0u8; 32].

Tuple Fields of Address

0: H160

FixedBytes(Vec<u8, Global>)

Vector of bytes with known size.

solidity name eg.: bytes8, bytes32, bytes64, bytes1024 Encoded to right padded [0u8; ((N + 31) / 32) * 32].

Tuple Fields of FixedBytes

0: Vec<u8, Global>

Bytes(Vec<u8, Global>)

Vector of bytes of unknown size.

solidity name: bytes Encoded in two parts. Init part: offset of ’closing part`. Closing part: encoded length followed by encoded right padded bytes.

Tuple Fields of Bytes

0: Vec<u8, Global>

Int(U256)

Signed integer.

solidity name: int

Tuple Fields of Int

0: U256

Uint(U256)

Unisnged integer.

solidity name: uint

Tuple Fields of Uint

0: U256

Bool(bool)

Boolean value.

solidity name: bool Encoded as left padded [0u8; 32], where last bit represents boolean value.

Tuple Fields of Bool

0: bool

String(String)

String.

solidity name: string Encoded in the same way as bytes. Must be utf8 compliant.

Tuple Fields of String

0: String

FixedArray(Vec<Token, Global>)

Array with known size.

solidity name eg.: int[3], bool[3], address[][8] Encoding of array is equal to encoding of consecutive elements of array.

Tuple Fields of FixedArray

0: Vec<Token, Global>

Array(Vec<Token, Global>)

Array of params with unknown size.

solidity name eg. int[], bool[], address[5][]

Tuple Fields of Array

0: Vec<Token, Global>

Tuple(Vec<Token, Global>)

Tuple of params of variable types.

solidity name: tuple

Tuple Fields of Tuple

0: Vec<Token, Global>

Implementations

impl Token

pub fn type_check(&self, param_type: &ParamType) -> bool

Check whether the type of the token matches the given parameter type.

Numeric types (Int and Uint) type check if the size of the token type is of greater or equal size than the provided parameter type.

pub fn into_address(self) -> Option<H160>

Converts token to…

pub fn into_fixed_bytes(self) -> Option<Vec<u8, Global>>

Converts token to…

pub fn into_bytes(self) -> Option<Vec<u8, Global>>

Converts token to…

pub fn into_int(self) -> Option<U256>

Converts token to…

pub fn into_uint(self) -> Option<U256>

Converts token to…

pub fn into_bool(self) -> Option<bool>

Converts token to…

pub fn into_string(self) -> Option<String>

Converts token to…

pub fn into_fixed_array(self) -> Option<Vec<Token, Global>>

Converts token to…

pub fn into_array(self) -> Option<Vec<Token, Global>>

Converts token to…

pub fn types_check(tokens: &[Token], param_types: &[ParamType]) -> bool

Check if all the types of the tokens match the given parameter types.

pub fn is_dynamic(&self) -> bool

Check if the token is a dynamic type resulting in prefixed encoding