ethers_core::abi
Trait AbiArrayType

pub trait AbiArrayType: AbiType { }

Additional trait for types that can appear in arrays

NOTE: this is necessary to handle the special case of Vec<u8> => Bytes

Object Safety

This trait is not object safe.