ethers_core::abi
Trait AbiType

pub trait AbiType {
    // Required method
    fn param_type() -> ParamType;

    // Provided method
    fn minimum_size() -> usize { ... }
}

A trait for types that can be represented in the Ethereum ABI.

Required Methods

fn param_type() -> ParamType

The native ABI type this type represents.

Provided Methods

fn minimum_size() -> usize

A hint of the minimum number of bytes this type takes up in the ABI.

Object Safety

This trait is not object safe.