ethers_core
Module types

Ethereum data types.

Re-exports

pub use transaction::eip1559::Eip1559TransactionRequest;

pub use transaction::eip2930::Eip2930TransactionRequest;

pub use transaction::request::TransactionRequest;

pub use transaction::response::Transaction;

pub use transaction::response::TransactionReceipt;

Modules

serde_helpers

Some convenient serde helpers

spoof

Provides types and methods for constructing an eth_call state override set

transaction

Structs

AccountDiff

Serde-friendly AccountDiff shadow.

AccountState

Block

The block type returned from RPC calls.

BlockOverrides

Bindings for block overrides in debug_traceCall options

BlockTrace

Ad-Hoc trace API type

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

Bytes

Wrapper type around Bytes to deserialize/serialize “0x” prefixed ethereum hex strings

Call

Call response

CallConfig

CallFrame

CallLogFrame

CallResult

Call Result

ChainIter

An iterator over the variants of Chain

ChangedType

Aux type for Diff::Changed.

Create

Create response

CreateResult

Create Result

DefaultFrame

DiffMode

EIP1186ProofResponse

FeeHistory

Filter

Filter for

FilteredParams

Support for matching Filters

FourByteFrame

GethDebugTracingCallOptions

Bindings for additional debug_traceCall options

GethDebugTracingOptions

Bindings for additional debug_traceTransaction options

H32

H64

H128

Fixed-size uninterpreted hash type with 16 bytes (128 bits) size.

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

I256

Little-endian 256-bit signed integer.

Log

A log produced by a transaction.

MemoryDiff

A diff of some chunk of memory.

NoopFrame

OtherFields

A type that is supposed to capture additional fields that are not native to ethereum but included in ethereum adjacent networks, for example fields the optimism eth_getTransactionByHash request returns additional fields that this type will capture

ParseBytesError

PreStateConfig

PreStateMode

Reward

Reward action

Signature

An ECDSA signature

StateDiff

Serde-friendly StateDiff shadow.

StorageDiff

A diff of some storage value.

StorageProof

StructLog

Suicide

Suicide

SyncProgress

Represents the sync status of the node

Trace

Trace-Filtering API trace type

TraceFilter

Trace filter

TransactionTrace

Trace

TxHash

A transaction Hash Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

TxpoolContent

Transaction Pool Content

TxpoolInspect

Transaction Pool Inspect

TxpoolInspectSummary

Transaction summary as found in the Txpool Inspection property.

TxpoolStatus

Transaction Pool Status

U64

Little-endian large integer type Unsigned 64-bit integer.

U128

Little-endian large integer type 128-bit unsigned integer.

U256

Little-endian large integer type 256-bit unsigned integer.

U512

Little-endian large integer type 512-bits unsigned integer.

Uint8

A wrapper for u8

VMExecutedOperation

A record of an executed VM operation.

VMOperation

A record of the execution of a single VM operation.

VMTrace

A record of a full VM trace for a CALL/CREATE.

Withdrawal

A validator withdrawal from the consensus layer. See EIP-4895: Beacon chain push withdrawals as operations.

Enums

Action

Action

ActionType

An external action type.

AddressOrBytes

A type that can either be an Address or Bytes.

BlockId

A block hash or block number.

BlockNumber

A block number or tag.

CallType

Call type.

Chain

An Ethereum EIP-155 chain.

Diff

Serde-friendly Diff shadow.

ExecutedInstruction

Helper to classify the executed instruction

FilterBlockOption

Represents the target range of blocks for the filter

GethDebugBuiltInTracerConfig

GethDebugBuiltInTracerType

Available built-in tracers

GethDebugTracerConfig

GethDebugTracerType

Available tracers

GethTrace

GethTraceFrame

NameOrAddress

ENS name or Ethereum Address. Not RLP encoded/serialized if it’s a name.

Opcode

An EVM Opcode.

ParseI256Error

The error type that is returned when parsing a 256-bit signed integer.

PathOrString

A type that can either be a Path or a String

PreStateFrame

RecoveryMessage

Recovery message data.

Res

Response

RewardType

Reward type.

Sign

Enum to represent the sign of a 256-bit signed integer.

SignatureError

An error involving a signature.

SyncingStatus

Structure used in eth_syncing RPC

TimeError

Error returned by Block::time.

TraceError

Trace evm errors.

TraceType

Description of the type of trace to make

ValueOrArray

Union type for representing a single value or a vector of values inside a filter

Traits

BigEndianHash

Functions

deserialize_bytes

serialize_bytes

u256_from_f64_saturating

Convert a floating point value to its nearest f64 integer.

Type Aliases

Address

BloomFilter

Selector

Topic

A single topic