Enum ethers_core::types::Opcode

#[repr(u8)]
pub enum Opcode {

    STOP,
    ADD,
    MUL,
    SUB,
    DIV,
    SDIV,
    MOD,
    SMOD,
    ADDMOD,
    MULMOD,
    EXP,
    SIGNEXTEND,
    LT,
    GT,
    SLT,
    SGT,
    EQ,
    ISZERO,
    AND,
    OR,
    XOR,
    NOT,
    BYTE,
    SHL,
    SHR,
    SAR,
    SHA3,
    ADDRESS,
    BALANCE,
    ORIGIN,
    CALLER,
    CALLVALUE,
    CALLDATALOAD,
    CALLDATASIZE,
    CALLDATACOPY,
    CODESIZE,
    CODECOPY,
    GASPRICE,
    EXTCODESIZE,
    EXTCODECOPY,
    RETURNDATASIZE,
    RETURNDATACOPY,
    EXTCODEHASH,
    BLOCKHASH,
    COINBASE,
    TIMESTAMP,
    NUMBER,
    DIFFICULTY,
    GASLIMIT,
    CHAINID,
    SELFBALANCE,
    BASEFEE,
    POP,
    MLOAD,
    MSTORE,
    MSTORE8,
    SLOAD,
    SSTORE,
    JUMP,
    JUMPI,
    PC,
    MSIZE,
    GAS,
    JUMPDEST,
    PUSH0,
    PUSH1,
    PUSH2,
    PUSH3,
    PUSH4,
    PUSH5,
    PUSH6,
    PUSH7,
    PUSH8,
    PUSH9,
    PUSH10,
    PUSH11,
    PUSH12,
    PUSH13,
    PUSH14,
    PUSH15,
    PUSH16,
    PUSH17,
    PUSH18,
    PUSH19,
    PUSH20,
    PUSH21,
    PUSH22,
    PUSH23,
    PUSH24,
    PUSH25,
    PUSH26,
    PUSH27,
    PUSH28,
    PUSH29,
    PUSH30,
    PUSH31,
    PUSH32,
    DUP1,
    DUP2,
    DUP3,
    DUP4,
    DUP5,
    DUP6,
    DUP7,
    DUP8,
    DUP9,
    DUP10,
    DUP11,
    DUP12,
    DUP13,
    DUP14,
    DUP15,
    DUP16,
    SWAP1,
    SWAP2,
    SWAP3,
    SWAP4,
    SWAP5,
    SWAP6,
    SWAP7,
    SWAP8,
    SWAP9,
    SWAP10,
    SWAP11,
    SWAP12,
    SWAP13,
    SWAP14,
    SWAP15,
    SWAP16,
    LOG0,
    LOG1,
    LOG2,
    LOG3,
    LOG4,
    CREATE,
    CALL,
    CALLCODE,
    RETURN,
    DELEGATECALL,
    CREATE2,
    STATICCALL,
    REVERT,
    INVALID,
    SELFDESTRUCT,

}

An EVM Opcode.

Variants

STOP

Opcode 0x0 - Halts execution

ADD

Opcode 0x1 - Addition operation

MUL

Opcode 0x2 - Multiplication operation

SUB

Opcode 0x3 - Subtraction operation

DIV

Opcode 0x4 - Integer division operation

SDIV

Opcode 0x5 - Signed integer division operation (truncated)

MOD

Opcode 0x6 - Modulo remainder operation

SMOD

Opcode 0x7 - Signed modulo remainder operation

ADDMOD

Opcode 0x8 - Modulo addition operation

MULMOD

Opcode 0x9 - Modulo multiplication operation

EXP

Opcode 0xA - Exponential operation

SIGNEXTEND

Opcode 0xB - Extend length of two’s complement signed integer

LT

Opcode 0x10 - Less-than comparison

GT

Opcode 0x11 - Greater-than comparison

SLT

Opcode 0x12 - Signed less-than comparison

SGT

Opcode 0x13 - Signed greater-than comparison

EQ

Opcode 0x14 - Equality comparison

ISZERO

Opcode 0x15 - Simple not operator

AND

Opcode 0x16 - Bitwise AND operation

OR

Opcode 0x17 - Bitwise OR operation

XOR

Opcode 0x18 - Bitwise XOR operation

NOT

Opcode 0x19 - Bitwise NOT operation

BYTE

Opcode 0x1A - Retrieve single byte from word

SHL

Opcode 0x1B - Left shift operation

SHR

Opcode 0x1C - Logical right shift operation

SAR

Opcode 0x1D - Arithmetic (signed) right shift operation

SHA3

Opcode 0x20 - Compute Keccak-256 hash

ADDRESS

Opcode 0x30 - Get address of currently executing account

BALANCE

Opcode 0x31 - Get address of currently executing account

ORIGIN

Opcode 0x32 - Get execution origination address

CALLER

Opcode 0x33 - Get caller address

CALLVALUE

Opcode 0x34 - Get deposited value by the instruction/transaction responsible for this execution

CALLDATALOAD

Opcode 0x35 - Get input data of current environment

CALLDATASIZE

Opcode 0x36 - Get size of input data in current environment

CALLDATACOPY

Opcode 0x37 - Copy input data in current environment to memory

CODESIZE

Opcode 0x38 - Get size of code running in current environment

CODECOPY

Opcode 0x39 - Copy code running in current environment to memory

GASPRICE

Opcode 0x3A - Get price of gas in current environment

EXTCODESIZE

Opcode 0x3B - Get size of an account’s code

EXTCODECOPY

Opcode 0x3C - Copy an account’s code to memory

RETURNDATASIZE

Opcode 0x3D - Get size of output data from the previous call from the current environment

RETURNDATACOPY

Opcode 0x3E - Copy output data from the previous call to memory

EXTCODEHASH

Opcode 0x3F - Get hash of an account’s code

BLOCKHASH

Opcode 0x40 - Get the hash of one of the 256 most recent complete blocks

COINBASE

Opcode 0x41 - Get the block’s beneficiary address

TIMESTAMP

Opcode 0x42 - Get the block’s timestamp

NUMBER

Opcode 0x43 - Get the block’s number

DIFFICULTY

Opcode 0x44 - Get the block’s difficulty

GASLIMIT

Opcode 0x45 - Get the block’s gas limit

CHAINID

Opcode 0x46 - Get the chain ID

SELFBALANCE

Opcode 0x47 - Get balance of currently executing account

BASEFEE

Opcode 0x48 - Get the base fee

POP

Opcode 0x50 - Remove item from stack

MLOAD

Opcode 0x51 - Load word from memory

MSTORE

Opcode 0x52 - Save word to memory

MSTORE8

Opcode 0x53 - Save byte to memory

SLOAD

Opcode 0x54 - Load word from storage

SSTORE

Opcode 0x55 - Save word to storage

JUMP

Opcode 0x56 - Alter the program counter

JUMPI

Opcode 0x57 - Conditionally alter the program counter

PC

Opcode 0x58 - Get the value of the program counter prior to the increment corresponding to this instruction

MSIZE

Opcode 0x59 - Get the size of active memory in bytes

GAS

Opcode 0x5A - Get the amount of available gas, including the corresponding reduction for the cost of this instruction

JUMPDEST

Opcode 0x5B - Mark a valid destination for jumps

PUSH0

Opcode 0x5F - Place the constant value 0 on stack

PUSH1

Opcode 0x60 - Place 1 byte item on stack

PUSH2

Opcode 0x61 - Place 2 byte item on stack

PUSH3

Opcode 0x62 - Place 3 byte item on stack

PUSH4

Opcode 0x63 - Place 4 byte item on stack

PUSH5

Opcode 0x64 - Place 5 byte item on stack

PUSH6

Opcode 0x65 - Place 6 byte item on stack

PUSH7

Opcode 0x66 - Place 7 byte item on stack

PUSH8

Opcode 0x67 - Place 8 byte item on stack

PUSH9

Opcode 0x68 - Place 9 byte item on stack

PUSH10

Opcode 0x69 - Place 10 byte item on stack

PUSH11

Opcode 0x6A - Place 11 byte item on stack

PUSH12

Opcode 0x6B - Place 12 byte item on stack

PUSH13

Opcode 0x6C - Place 13 byte item on stack

PUSH14

Opcode 0x6D - Place 14 byte item on stack

PUSH15

Opcode 0x6E - Place 15 byte item on stack

PUSH16

Opcode 0x6F - Place 16 byte item on stack

PUSH17

Opcode 0x70 - Place 17 byte item on stack

PUSH18

Opcode 0x71 - Place 18 byte item on stack

PUSH19

Opcode 0x72 - Place 19 byte item on stack

PUSH20

Opcode 0x73 - Place 20 byte item on stack

PUSH21

Opcode 0x74 - Place 21 byte item on stack

PUSH22

Opcode 0x75 - Place 22 byte item on stack

PUSH23

Opcode 0x76 - Place 23 byte item on stack

PUSH24

Opcode 0x77 - Place 24 byte item on stack

PUSH25

Opcode 0x78 - Place 25 byte item on stack

PUSH26

Opcode 0x79 - Place 26 byte item on stack

PUSH27

Opcode 0x7A - Place 27 byte item on stack

PUSH28

Opcode 0x7B - Place 28 byte item on stack

PUSH29

Opcode 0x7C - Place 29 byte item on stack

PUSH30

Opcode 0x7D - Place 30 byte item on stack

PUSH31

Opcode 0x7E - Place 31 byte item on stack

PUSH32

Opcode 0x7F - Place 32 byte item on stack

DUP1

Opcode 0x80 - Duplicate 1st stack item

DUP2

Opcode 0x81 - Duplicate 2nd stack item

DUP3

Opcode 0x82 - Duplicate 3rd stack item

DUP4

Opcode 0x83 - Duplicate 4th stack item

DUP5

Opcode 0x84 - Duplicate 5th stack item

DUP6

Opcode 0x85 - Duplicate 6th stack item

DUP7

Opcode 0x86 - Duplicate 7th stack item

DUP8

Opcode 0x87 - Duplicate 8th stack item

DUP9

Opcode 0x88 - Duplicate 9th stack item

DUP10

Opcode 0x89 - Duplicate 10th stack item

DUP11

Opcode 0x8A - Duplicate 11th stack item

DUP12

Opcode 0x8B - Duplicate 12th stack item

DUP13

Opcode 0x8C - Duplicate 13th stack item

DUP14

Opcode 0x8D - Duplicate 14th stack item

DUP15

Opcode 0x8E - Duplicate 15th stack item

DUP16

Opcode 0x8F - Duplicate 16th stack item

SWAP1

Opcode 0x90 - Exchange 1st and 2nd stack items

SWAP2

Opcode 0x91 - Exchange 1st and 3rd stack items

SWAP3

Opcode 0x92 - Exchange 1st and 4th stack items

SWAP4

Opcode 0x93 - Exchange 1st and 5th stack items

SWAP5

Opcode 0x94 - Exchange 1st and 6th stack items

SWAP6

Opcode 0x95 - Exchange 1st and 7th stack items

SWAP7

Opcode 0x96 - Exchange 1st and 8th stack items

SWAP8

Opcode 0x97 - Exchange 1st and 9th stack items

SWAP9

Opcode 0x98 - Exchange 1st and 10th stack items

SWAP10

Opcode 0x99 - Exchange 1st and 11th stack items

SWAP11

Opcode 0x9A - Exchange 1st and 12th stack items

SWAP12

Opcode 0x9B - Exchange 1st and 13th stack items

SWAP13

Opcode 0x9C - Exchange 1st and 14th stack items

SWAP14

Opcode 0x9D - Exchange 1st and 15th stack items

SWAP15

Opcode 0x9E - Exchange 1st and 16th stack items

SWAP16

Opcode 0x9F - Exchange 1st and 17th stack items

LOG0

Opcode 0xA0 - Append log record with one topic

LOG1

Opcode 0xA1 - Append log record with two topics

LOG2

Opcode 0xA2 - Append log record with three topics

LOG3

Opcode 0xA3 - Append log record with four topics

LOG4

Opcode 0xA4 - Append log record with five topics

CREATE

Opcode 0xF0 - Create a new account with associated code

CALL

Opcode 0xF1 - Message-call into an account

CALLCODE

Opcode 0xF2 - Message-call into this account with alternative account’s code

RETURN

Opcode 0xF3 - Halt execution returning output data

DELEGATECALL

Opcode 0xF4 - Message-call into this account with an alternative account’s code, but persisting the current values for sender and value

CREATE2

Opcode 0xF5 - Create a new account with associated code at a predictable address

STATICCALL

Opcode 0xFA - Static message-call into an account

REVERT

Opcode 0xFD - Halt execution reverting state changes but returning data and remaining gas

INVALID

Opcode 0xFE - Designated invalid instruction

SELFDESTRUCT

Opcode 0xFF - Halt execution and register account for later deletion