Enum ethers_core::types::TraceError

#[repr(u8)]
pub enum TraceError {

    OutOfGas = 0,
    BadJumpDestination = 1,
    BadInstruction = 2,
    StackUnderflow = 3,
    OutOfStack = 4,
    Internal = 5,
    BuiltIn = 6,
    MutableCallInStaticContext = 7,
    Wasm = 8,
    OutOfBounds = 9,
    Reverted = 10,
    SubStackUnderflow = 11,
    OutOfSubStack = 12,
    InvalidSubEntry = 13,
    InvalidCode = 14,

}

Trace evm errors.

Variants

OutOfGas = 0

OutOfGas is returned when transaction execution runs out of gas.

BadJumpDestination = 1

BadJumpDestination is returned when execution tried to move to position that wasn’t marked with JUMPDEST instruction

BadInstruction = 2

BadInstructions is returned when given instruction is not supported

StackUnderflow = 3

StackUnderflow when there is not enough stack elements to execute instruction

OutOfStack = 4

When execution would exceed defined Stack Limit

Internal = 5

Returned on evm internal error. Should never be ignored during development. Likely to cause consensus issues.

BuiltIn = 6

When builtin contract failed on input data

MutableCallInStaticContext = 7

When execution tries to modify the state in static context

Wasm = 8

Wasm error

OutOfBounds = 9

Contract tried to access past the return data buffer.

Reverted = 10

Execution has been reverted with REVERT instruction.

SubStackUnderflow = 11

When there is not enough subroutine stack elements to return from

OutOfSubStack = 12

When execution would exceed defined subroutine Stack Limit

InvalidSubEntry = 13

When the code walks into a subroutine, that is not allowed

InvalidCode = 14

When invalid code was attempted to deploy

Implementations

impl TraceError

pub fn from_u8(val: u8) -> Result<Self, u8>

Converts the given byte into the corresponding error variant.