Struct ethers_core::types::Block

pub struct Block<TX> {

    pub hash: Option<H256>,
    pub parent_hash: H256,
    pub uncles_hash: H256,
    pub author: Option<Address>,
    pub state_root: H256,
    pub transactions_root: H256,
    pub receipts_root: H256,
    pub number: Option<U64>,
    pub gas_used: U256,
    pub gas_limit: U256,
    pub extra_data: Bytes,
    pub logs_bloom: Option<Bloom>,
    pub timestamp: U256,
    pub difficulty: U256,
    pub total_difficulty: Option<U256>,
    pub seal_fields: Vec<Bytes>,
    pub uncles: Vec<H256>,
    pub transactions: Vec<TX>,
    pub size: Option<U256>,
    pub mix_hash: Option<H256>,
    pub nonce: Option<H64>,
    pub base_fee_per_gas: Option<U256>,
    pub blob_gas_used: Option<U256>,
    pub excess_blob_gas: Option<U256>,
    pub withdrawals_root: Option<H256>,
    pub withdrawals: Option<Vec<Withdrawal>>,
    pub parent_beacon_block_root: Option<H256>,
    pub other: OtherFields,

}

The block type returned from RPC calls.

This is generic over a TX type which will be either the hash or the full transaction, i.e. Block<TxHash> or Block<Transaction>.

Fields
hash: Option<H256>

Hash of the block

parent_hash: H256

Hash of the parent

uncles_hash: H256

Hash of the uncles

author: Option<Address>

Miner/author’s address. None if pending.

state_root: H256

State root hash

transactions_root: H256

Transactions root hash

receipts_root: H256

Transactions receipts root hash

number: Option<U64>

Block number. None if pending.

gas_used: U256

Gas Used

gas_limit: U256

Gas Limit

extra_data: Bytes

Extra data

logs_bloom: Option<Bloom>

Logs bloom

timestamp: U256

Timestamp

difficulty: U256

Difficulty

total_difficulty: Option<U256>

Total difficulty

seal_fields: Vec<Bytes>

Seal fields

uncles: Vec<H256>

Uncles’ hashes

transactions: Vec<TX>

Transactions

size: Option<U256>

Size in bytes

mix_hash: Option<H256>

Mix Hash

nonce: Option<H64>

Nonce

base_fee_per_gas: Option<U256>

Base fee per unit of gas (if past London)

blob_gas_used: Option<U256>

Blob gas used (if past Cancun)

excess_blob_gas: Option<U256>

Excess blob gas (if past Cancun)

withdrawals_root: Option<H256>

Withdrawals root hash (if past Shanghai)

withdrawals: Option<Vec<Withdrawal>>

Withdrawals (if past Shanghai)

parent_beacon_block_root: Option<H256>

Parent beacon block root (if past Cancun)

other: OtherFields

Captures unknown fields such as additional fields used by L2s

Implementations

impl<TX> Block<TX>

pub fn gas_target(&self) -> U256

The target gas usage as per EIP-1559

pub fn next_block_base_fee(&self) -> Option<U256>

The next block’s base fee, it is a function of parent block’s base fee and gas usage. Reference: https://eips.ethereum.org/EIPS/eip-1559

pub fn time(&self) -> Result<DateTime<Utc>, TimeError>

Parse Self::timestamp into a DateTime<Utc>.

Errors

TimeError::TimestampZero if the timestamp is zero, or

TimeError::TimestampOverflow if the timestamp is too large to be represented as a DateTime<Utc>.

impl Block<TxHash>

pub fn into_full_block( self, transactions: Vec<Transaction> ) -> Block<Transaction>

Converts this block that only holds transaction hashes into a full block with Transaction