Struct ethers_core::types::Filter

pub struct Filter {
    pub block_option: FilterBlockOption,
    pub address: Option<ValueOrArray<Address>>,
    pub topics: [Option<Topic>; 4],
}

Filter for

Fields
block_option: FilterBlockOption

Filter block options, specifying on which blocks the filter should match.

address: Option<ValueOrArray<Address>>

Address

topics: [Option<Topic>; 4]

Topics

Implementations

impl Filter

pub fn new() -> Self

pub fn select(self, filter: impl Into<FilterBlockOption>) -> Self

Sets the inner filter object

NOTE: ranges are always inclusive

Examples

Match only a specific block

let filter = Filter::new().select(69u64);

This is the same as Filter::new().from_block(1337u64).to_block(1337u64)

Match the latest block only

let filter = Filter::new().select(BlockNumber::Latest);

Match a block by its hash

let filter = Filter::new().select(H256::zero());

This is the same as at_block_hash

Match a range of blocks

let filter = Filter::new().select(0u64..100u64);

Match all blocks in range (1337..BlockNumber::Latest)

let filter = Filter::new().select(1337u64..);

Match all blocks in range (BlockNumber::Earliest..1337)

let filter = Filter::new().select(..1337u64);

pub fn from_block<T: Into<BlockNumber>>(self, block: T) -> Self

pub fn to_block<T: Into<BlockNumber>>(self, block: T) -> Self

pub fn at_block_hash<T: Into<H256>>(self, hash: T) -> Self

pub fn address<T: Into<ValueOrArray<Address>>>(self, address: T) -> Self

Sets the inner filter object

NOTE: ranges are always inclusive

Examples

Match only a specific address ("0xAc4b3DacB91461209Ae9d41EC517c2B9Cb1B7DAF")

let filter = Filter::new().address("0xAc4b3DacB91461209Ae9d41EC517c2B9Cb1B7DAF".parse::<Address>().unwrap());

Match all addresses in array (vec!["0xAc4b3DacB91461209Ae9d41EC517c2B9Cb1B7DAF", "0x8ad599c3A0ff1De082011EFDDc58f1908eb6e6D8"])

let addresses = vec!["0xAc4b3DacB91461209Ae9d41EC517c2B9Cb1B7DAF".parse::<Address>().unwrap(),"0x8ad599c3A0ff1De082011EFDDc58f1908eb6e6D8".parse::<Address>().unwrap()];
let filter = Filter::new().address(addresses);

pub fn event(self, event_name: &str) -> Self

Given the event signature in string form, it hashes it and adds it to the topics to monitor

pub fn events(self, events: impl IntoIterator<Item = impl AsRef<[u8]>>) -> Self

Hashes all event signatures and sets them as array to topic0

pub fn topic0<T: Into<Topic>>(self, topic: T) -> Self

Sets topic0 (the event name for non-anonymous events)

pub fn topic1<T: Into<Topic>>(self, topic: T) -> Self

Sets the 1st indexed topic

pub fn topic2<T: Into<Topic>>(self, topic: T) -> Self

Sets the 2nd indexed topic

pub fn topic3<T: Into<Topic>>(self, topic: T) -> Self

Sets the 3rd indexed topic

pub fn is_paginatable(&self) -> bool

pub fn get_to_block(&self) -> Option<U64>

Returns the numeric value of the toBlock field

pub fn get_from_block(&self) -> Option<U64>

Returns the numeric value of the fromBlock field

pub fn get_block_hash(&self) -> Option<H256>

Returns the numeric value of the fromBlock field

pub fn topics(&self) -> impl Iterator<Item = &Topic> + '_

Returns an iterator over all existing topics

pub fn has_topics(&self) -> bool

Returns true if at least one topic is set