Struct ethers_core::types::I256

pub struct I256(_);

Little-endian 256-bit signed integer.

Implementations

impl I256

pub const MAX: I256

Maximum value.

pub const MIN: I256

Minimum value.

pub const fn zero() -> Self

Zero (additive identity) of this type.

pub const fn one() -> Self

One (multiplicative identity) of this type.

pub const fn minus_one() -> Self

Minus one (multiplicative inverse) of this type.

pub const fn max_value() -> Self

The maximum value which can be inhabited by this type.

pub const fn min_value() -> Self

The minimum value which can be inhabited by this type.

pub fn overflowing_from_sign_and_abs(sign: Sign, abs: U256) -> (Self, bool)

Creates an I256 from a sign and an absolute value. Returns the value and a bool that is true if the conversion caused an overflow.

pub fn checked_from_sign_and_abs(sign: Sign, abs: U256) -> Option<Self>

Creates an I256 from an absolute value and a negative flag. Returns None if it would overflow an I256.

pub fn into_sign_and_abs(self) -> (Sign, U256)

Splits a I256 into its absolute value and negative flag.

pub fn sign(self) -> Sign

Returns the sign of self.

pub const fn from_raw(raw: U256) -> Self

Coerces an unsigned integer into a signed one. If the unsigned integer is greater than the greater than or equal to 1 << 255, then the result will overflow into a negative value.

pub fn into_raw(self) -> U256

Returns the signed integer as a unsigned integer. If the value of self negative, then the two’s complement of its absolute value will be returned.

pub fn low_i32(&self) -> i32

Conversion to i32

pub fn low_u32(&self) -> u32

Conversion to u32

pub fn low_i64(&self) -> i64

Conversion to i64

pub fn low_u64(&self) -> u64

Conversion to u64

pub fn low_i128(&self) -> i128

Conversion to i128

pub fn low_u128(&self) -> u128

Conversion to u128

pub fn low_isize(&self) -> isize

Conversion to i128

pub fn low_usize(&self) -> usize

Conversion to usize

pub fn as_i32(&self) -> i32

Conversion to i32 with overflow checking

Panics

Panics if the number is outside the range [i32::MIN, i32::MAX].

pub fn as_u32(&self) -> u32

Conversion to u32 with overflow checking

Panics

Panics if the number is outside the range [0, u32::MAX].

pub fn as_i64(&self) -> i64

Conversion to i64 with overflow checking

Panics

Panics if the number is outside the range [i64::MIN, i64::MAX].

pub fn as_u64(&self) -> u64

Conversion to u64 with overflow checking

Panics

Panics if the number is outside the range [0, u64::MAX].

pub fn as_i128(&self) -> i128

Conversion to i128 with overflow checking

Panics

Panics if the number is outside the range [i128::MIN, i128::MAX].

pub fn as_u128(&self) -> u128

Conversion to u128 with overflow checking

Panics

Panics if the number is outside the range [0, u128::MAX].

pub fn as_isize(&self) -> usize

Conversion to isize with overflow checking

Panics

Panics if the number is outside the range [isize::MIN, isize::MAX].

pub fn as_usize(&self) -> usize

Conversion to usize with overflow checking

Panics

Panics if the number is outside the range [0, usize::MAX].

pub fn from_dec_str(value: &str) -> Result<Self, ParseI256Error>

Convert from a decimal string.

pub fn from_hex_str(value: &str) -> Result<Self, ParseI256Error>

Convert from a hexadecimal string.

pub fn signum(self) -> Self

Returns the sign of the number.

pub fn is_positive(self) -> bool

Returns true if self is positive and false if the number is zero or negative.

pub fn is_negative(self) -> bool

Returns true if self is negative and false if the number is zero or negative.

pub fn is_zero(self) -> bool

Returns true if self is negative and false if the number is zero or positive.

pub fn abs(self) -> Self

Gets the absolute value.

Panics

In debug mode, will panic if it overflows.

pub fn overflowing_abs(self) -> (Self, bool)

Computes the absolute value of self.

Returns a tuple of the absolute version of self along with a boolean indicating whether an overflow happened. If self is the minimum value (e.g., I256::MIN for values of type I256), then the minimum value will be returned again and true will be returned for an overflow happening.

pub fn checked_abs(self) -> Option<Self>

Checked absolute value. Computes self.abs(), returning None if self == MIN.

pub fn saturating_abs(self) -> Self

Saturating absolute value. Computes self.abs(), returning MAX if self == MIN instead of overflowing.

pub fn wrapping_abs(self) -> Self

Wrapping absolute value. Computes self.abs(), wrapping around at the boundary of the type.

pub fn overflowing_neg(self) -> (Self, bool)

Negates self, overflowing if this is equal to the minimum value.

Returns a tuple of the negated version of self along with a boolean indicating whether an overflow happened. If self is the minimum value, then the minimum value will be returned again and true will be returned for an overflow happening.

pub fn checked_neg(self) -> Option<Self>

Checked negation. Computes self.neg(), returning None if self == MIN.

pub fn saturating_neg(self) -> Self

Saturating negation. Computes self.neg(), returning MAX if self == MIN instead of overflowing.

pub fn wrapping_neg(self) -> Self

Wrapping negation. Computes self.neg(), returning MIN if self == MIN instead of overflowing.

pub fn bits(&self) -> u32

Return the least number of bits needed to represent the number.

pub fn bit(&self, index: usize) -> bool

Return if specific bit is set.

Panics

Panics if index exceeds the bit width of the number.

pub fn count_ones(&self) -> u32

Returns the number of ones in the binary representation of self.

pub fn count_zeros(&self) -> u32

Returns the number of zeros in the binary representation of self.

pub fn leading_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn trailing_zeros(&self) -> u32

Returns the number of leading zeros in the binary representation of self.

pub fn byte(&self, index: usize) -> u8

Return specific byte.

Panics

Panics if index exceeds the byte width of the number.

pub fn to_big_endian(&self, bytes: &mut [u8])

Write to the slice in big-endian format.

pub fn to_little_endian(&self, bytes: &mut [u8])

Write to the slice in little-endian format.

pub fn overflowing_add(self, rhs: Self) -> (Self, bool)

Calculates self + rhs

Returns a tuple of the addition along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_add(self, other: Self) -> Option<Self>

Checked addition. Returns None if overflow occurred.

pub fn saturating_add(self, other: Self) -> Self

Addition which saturates at the maximum value (Self::max_value()).

pub fn wrapping_add(self, other: Self) -> Self

Wrapping addition.

pub fn overflowing_sub(self, rhs: Self) -> (Self, bool)

Calculates self - rhs

Returns a tuple of the subtraction along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_sub(self, other: Self) -> Option<Self>

Checked subtraction. Returns None if overflow occurred.

pub fn saturating_sub(self, other: Self) -> Self

Subtraction which saturates at zero.

pub fn wrapping_sub(self, other: Self) -> Self

Wrapping subtraction.

pub fn overflowing_mul(self, rhs: Self) -> (Self, bool)

Calculates self * rhs

Returns a tuple of the multiplication along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_mul(self, other: Self) -> Option<Self>

Checked multiplication. Returns None if overflow occurred.

pub fn saturating_mul(self, rhs: Self) -> Self

Multiplication which saturates at the maximum value..

pub fn wrapping_mul(self, rhs: Self) -> Self

Wrapping multiplication.

pub fn overflowing_div(self, rhs: Self) -> (Self, bool)

Calculates self / rhs

Returns a tuple of the division along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_div(self, rhs: Self) -> Option<Self>

Checked division. Returns None if overflow occurred or if rhs == 0.

pub fn saturating_div(self, rhs: Self) -> Self

Division which saturates at the maximum value.

pub fn wrapping_div(self, rhs: Self) -> Self

Wrapping division.

pub fn overflowing_rem(self, rhs: Self) -> (Self, bool)

Calculates self % rhs

Returns a tuple of the remainder along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would have occurred then the wrapped value is returned.

pub fn checked_rem(self, rhs: Self) -> Option<Self>

Checked remainder. Returns None if overflow occurred or rhs == 0

pub fn wrapping_rem(self, rhs: Self) -> Self

Wrapping remainder. Returns the result of the operation % regardless of whether or not the division overflowed.

pub fn div_euclid(self, rhs: Self) -> Self

Calculates the quotient of Euclidean division of self by rhs.

This computes the integer n such that self = n * rhs + self.rem_euclid(rhs), with 0 <= self.rem_euclid(rhs) < rhs. In other words, the result is self / rhs rounded to the integer n such that self >= n * rhs:

If self > 0, this is equal to round towards zero (the default in Rust);

If self < 0, this is equal to round towards +/- infinity.

pub fn rem_euclid(self, rhs: Self) -> Self

Calculates the least non-negative remainder of self (mod rhs). This is done as if by the Euclidean division algorithm given r = self.rem_euclid(rhs), self = rhs * self.div_euclid(rhs) + r, and 0 <= r < abs(rhs).

pub fn overflowing_div_euclid(self, rhs: Self) -> (Self, bool)

Calculates the quotient of Euclidean division self.div_euclid(rhs). Returns a tuple of the divisor along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then self is returned.

pub fn checked_div_euclid(self, rhs: Self) -> Option<Self>

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn wrapping_div_euclid(self, rhs: Self) -> Self

Wrapping Euclidean division. Computes self.div_euclid(rhs), wrapping around at the boundary of the type. Wrapping only occurs in MIN / -1 on a signed type (where MIN is the negative minimal value for the type). This is equivalent to -MIN, a positive value that is too large to represent in the type. In this case, this method returns MIN itself.

pub fn overflowing_rem_euclid(self, rhs: Self) -> (Self, bool)

Overflowing Euclidean remainder. Calculates self.rem_euclid(rhs). Returns a tuple of the remainder after dividing along with a boolean indicating whether an arithmetic overflow would occur. If an overflow would occur then 0 is returned. Panics if rhs == 0

pub fn wrapping_rem_euclid(self, rhs: Self) -> Self

Wrapping Euclidean remainder. Computes self.rem_euclid(rhs), wrapping around at the boundary of the type. Wrapping will only occur in MIN % -1 on a signed type (where MIN is the negative minimal value for the type). In this case, this method returns 0. Panics when rhs == 0

pub fn checked_rem_euclid(self, rhs: Self) -> Option<Self>

Checked Euclidean remainder. Computes self.rem_euclid(rhs), returning None if rhs == 0 or the division results in overflow.

pub fn exp10(n: usize) -> Self

Create 10**n as this type.

Panics

Panics if the result overflows the type.

pub fn pow(self, exp: u32) -> Self

Raise self to the power of exp.

Panics

Panics if the result overflows the type in debug mode.

pub fn overflowing_pow(self, exp: u32) -> (Self, bool)

Raises self to the power of exp.

Returns a tuple of the exponentiation along with a bool indicating whether an overflow happened.

pub fn checked_pow(self, exp: u32) -> Option<Self>

Raises self to the power of exp. Returns None if overflow occurred.

pub fn saturating_pow(self, exp: u32) -> Self

Raises self to the power of exp, saturating at the numeric bounds instead of overflowing.

pub fn wrapping_pow(self, exp: u32) -> Self

Wrapping powolute value. Computes self.pow(), wrapping around at the boundary of the type.