Struct ethers_core::types::transaction::eip1559::Eip1559TransactionRequest

pub struct Eip1559TransactionRequest {
    pub from: Option<Address>,
    pub to: Option<NameOrAddress>,
    pub gas: Option<U256>,
    pub value: Option<U256>,
    pub data: Option<Bytes>,
    pub nonce: Option<U256>,
    pub access_list: AccessList,
    pub max_priority_fee_per_gas: Option<U256>,
    pub max_fee_per_gas: Option<U256>,
    pub chain_id: Option<U64>,
}

Parameters for sending a transaction

Fields
from: Option<Address>

Sender address or ENS name

to: Option<NameOrAddress>

Recipient address (None for contract creation)

gas: Option<U256>

Supplied gas (None for sensible default)

value: Option<U256>

Transferred value (None for no transfer)

data: Option<Bytes>

The compiled code of a contract OR the first 4 bytes of the hash of the invoked method signature and encoded parameters. For details see Ethereum Contract ABI

nonce: Option<U256>

Transaction nonce (None for next available nonce)

access_list: AccessListmax_priority_fee_per_gas: Option<U256>

Represents the maximum tx fee that will go to the miner as part of the user’s fee payment. It serves 3 purposes:

Compensates miners for the uncle/ommer risk + fixed costs of including transaction in a block;

Allows users with high opportunity costs to pay a premium to miners;

In times where demand exceeds the available block space (i.e. 100% full, 30mm gas), this component allows first price auctions (i.e. the pre-1559 fee model) to happen on the priority fee.

More context here

max_fee_per_gas: Option<U256>

Represents the maximum amount that a user is willing to pay for their tx (inclusive of baseFeePerGas and maxPriorityFeePerGas). The difference between maxFeePerGas and baseFeePerGas + maxPriorityFeePerGas is “refunded” to the user.

chain_id: Option<U64>

Chain ID (None for mainnet)

Implementations

impl Eip1559TransactionRequest

pub fn new() -> Self

Creates an empty transaction request with all fields left empty

pub fn from<T: Into<Address>>(self, from: T) -> Self

Sets the from field in the transaction to the provided value

pub fn to<T: Into<NameOrAddress>>(self, to: T) -> Self

Sets the to field in the transaction to the provided value

pub fn gas<T: Into<U256>>(self, gas: T) -> Self

Sets the gas field in the transaction to the provided value

pub fn max_priority_fee_per_gas<T: Into<U256>>( self, max_priority_fee_per_gas: T ) -> Self

Sets the max_priority_fee_per_gas field in the transaction to the provided value

pub fn max_fee_per_gas<T: Into<U256>>(self, max_fee_per_gas: T) -> Self

Sets the max_fee_per_gas field in the transaction to the provided value

pub fn value<T: Into<U256>>(self, value: T) -> Self

Sets the value field in the transaction to the provided value

pub fn data<T: Into<Bytes>>(self, data: T) -> Self

Sets the data field in the transaction to the provided value

pub fn access_list<T: Into<AccessList>>(self, access_list: T) -> Self

Sets the access_list field in the transaction to the provided value

pub fn nonce<T: Into<U256>>(self, nonce: T) -> Self

Sets the nonce field in the transaction to the provided value

pub fn chain_id<T: Into<U64>>(self, chain_id: T) -> Self

Sets the chain_id field in the transaction to the provided value

pub fn rlp(&self) -> Bytes

Gets the unsigned transaction’s RLP encoding

pub fn rlp_signed(&self, signature: &Signature) -> Bytes

Produces the RLP encoding of the transaction with the provided signature

pub fn decode_base_rlp( rlp: &Rlp<'_>, offset: &mut usize ) -> Result<Self, DecoderError>

Decodes fields of the request starting at the RLP offset passed. Increments the offset for each element parsed.

pub fn decode_signed_rlp( rlp: &Rlp<'_> ) -> Result<(Self, Signature), Eip1559RequestError>

Decodes the given RLP into a transaction, attempting to decode its signature as well.