Module ethers_core::types::transaction::eip712

Structs

EIP712Domain

Eip712 Domain attributes used in determining the domain separator; Unused fields are left out of the struct type.

EIP712WithDomain

Enums

Eip712Error

Error typed used by Eip712 derive macro

Constants

EIP712_DOMAIN_TYPE_HASH

Pre-computed value of the following statement:

EIP712_DOMAIN_TYPE_HASH_WITH_SALT

Pre-computed value of the following statement:

Traits

Eip712

The Eip712 trait provides helper methods for computing the typed data hash used in eth_signTypedData.

Functions

encode_eip712_type

Parse token into Eip712 compliant ABI encoding NOTE: Token::Tuple() is currently not supported for solidity structs; this is needed for nested Eip712 types, but is not implemented.

find_parameter_type

Parse the eth abi parameter type based on the syntax type; this method is copied from https://github.com/gakonst/ethers-rs/blob/master/ethers-contract/ethers-contract-derive/src/lib.rs#L600 with additional modifications for finding byte arrays

make_type_hash

Convert hash map of field names and types into a type hash corresponding to enc types;

parse_fields

Return HashMap of the field name and the field type;