Struct ethers_core::types::transaction::request::TransactionRequest

pub struct TransactionRequest {
    pub from: Option<Address>,
    pub to: Option<NameOrAddress>,
    pub gas: Option<U256>,
    pub gas_price: Option<U256>,
    pub value: Option<U256>,
    pub data: Option<Bytes>,
    pub nonce: Option<U256>,
    pub chain_id: Option<U64>,
}

Parameters for sending a transaction

Fields
from: Option<Address>

Sender address or ENS name

to: Option<NameOrAddress>

Recipient address (None for contract creation)

gas: Option<U256>

Supplied gas (None for sensible default)

gas_price: Option<U256>

Gas price (None for sensible default)

value: Option<U256>

Transferred value (None for no transfer)

data: Option<Bytes>

The compiled code of a contract OR the first 4 bytes of the hash of the invoked method signature and encoded parameters. For details see Ethereum Contract ABI

nonce: Option<U256>

Transaction nonce (None for next available nonce)

chain_id: Option<U64>

Chain ID (None for mainnet)

Implementations

impl TransactionRequest

pub fn new() -> Self

Creates an empty transaction request with all fields left empty

pub fn pay<T: Into<NameOrAddress>, V: Into<U256>>(to: T, value: V) -> Self

Convenience function for sending a new payment transaction to the receiver.

pub fn from<T: Into<Address>>(self, from: T) -> Self

Sets the from field in the transaction to the provided value

pub fn to<T: Into<NameOrAddress>>(self, to: T) -> Self

Sets the to field in the transaction to the provided value

pub fn gas<T: Into<U256>>(self, gas: T) -> Self

Sets the gas field in the transaction to the provided value

pub fn gas_price<T: Into<U256>>(self, gas_price: T) -> Self

Sets the gas_price field in the transaction to the provided value

pub fn value<T: Into<U256>>(self, value: T) -> Self

Sets the value field in the transaction to the provided value

pub fn data<T: Into<Bytes>>(self, data: T) -> Self

Sets the data field in the transaction to the provided value

pub fn nonce<T: Into<U256>>(self, nonce: T) -> Self

Sets the nonce field in the transaction to the provided value

pub fn chain_id<T: Into<U64>>(self, chain_id: T) -> Self

Sets the chain_id field in the transaction to the provided value

pub fn sighash(&self) -> H256

Hashes the transaction’s data with the provided chain id

pub fn rlp(&self) -> Bytes

Gets the transaction’s RLP encoding, prepared with the chain_id and extra fields for signing. Assumes the chainid exists.

pub fn rlp_unsigned(&self) -> Bytes

Gets the unsigned transaction’s RLP encoding

pub fn rlp_signed(&self, signature: &Signature) -> Bytes

Produces the RLP encoding of the transaction with the provided signature

pub fn decode_unsigned_rlp(rlp: &Rlp<'_>) -> Result<Self, DecoderError>

Decodes RLP into a transaction.

pub fn decode_signed_rlp( rlp: &Rlp<'_> ) -> Result<(Self, Signature), RequestError>

Decodes the given RLP into a transaction, attempting to decode its signature as well.

impl TransactionRequest

pub fn with_access_list<T: Into<AccessList>>( self, access_list: T ) -> Eip2930TransactionRequest

Sets the access_list field in the transaction (converts the TransactionRequest to an Eip2930TransactionRequest)