Struct ethers_core::types::transaction::response::Transaction

pub struct Transaction {

    pub hash: H256,
    pub nonce: U256,
    pub block_hash: Option<H256>,
    pub block_number: Option<U64>,
    pub transaction_index: Option<U64>,
    pub from: Address,
    pub to: Option<Address>,
    pub value: U256,
    pub gas_price: Option<U256>,
    pub gas: U256,
    pub input: Bytes,
    pub v: U64,
    pub r: U256,
    pub s: U256,
    pub transaction_type: Option<U64>,
    pub access_list: Option<AccessList>,
    pub max_priority_fee_per_gas: Option<U256>,
    pub max_fee_per_gas: Option<U256>,
    pub chain_id: Option<U256>,
    pub other: OtherFields,

}

Details of a signed transaction

Fields
hash: H256

The transaction’s hash

nonce: U256

The transaction’s nonce

block_hash: Option<H256>

Block hash. None when pending.

block_number: Option<U64>

Block number. None when pending.

transaction_index: Option<U64>

Transaction Index. None when pending.

from: Address

Sender

to: Option<Address>

Recipient (None when contract creation)

value: U256

Transferred value

gas_price: Option<U256>

Gas Price, null for Type 2 transactions

gas: U256

Gas amount

input: Bytes

Input data

v: U64

ECDSA recovery id

r: U256

ECDSA signature r

s: U256

ECDSA signature s

transaction_type: Option<U64>

Transaction type, Some(2) for EIP-1559 transaction, Some(1) for AccessList transaction, None for Legacy

access_list: Option<AccessList>max_priority_fee_per_gas: Option<U256>

Represents the maximum tx fee that will go to the miner as part of the user’s fee payment. It serves 3 purposes:

Compensates miners for the uncle/ommer risk + fixed costs of including transaction in a block; 2. Allows users with high opportunity costs to pay a premium to miners;

In times where demand exceeds the available block space (i.e. 100% full, 30mm gas), this component allows first price auctions (i.e. the pre-1559 fee model) to happen on the priority fee.

More context here

max_fee_per_gas: Option<U256>

Represents the maximum amount that a user is willing to pay for their tx (inclusive of baseFeePerGas and maxPriorityFeePerGas). The difference between maxFeePerGas and baseFeePerGas + maxPriorityFeePerGas is “refunded” to the user.

chain_id: Option<U256>other: OtherFields

Captures unknown fields such as additional fields used by L2s

Implementations

impl Transaction

pub fn hash(&self) -> H256

pub fn rlp(&self) -> Bytes

pub fn recover_from(&self) -> Result<Address, SignatureError>

Recover the sender of the tx from signature

pub fn recover_from_mut(&mut self) -> Result<Address, SignatureError>

Recover the sender of the tx from signature and set the from field