Module ethers_core::abi

This module implements extensions to the ethabi API.

Re-exports

pub use struct_def::SolStruct;

Modules

ethereum_types

param_type

Function and event param types.

struct_def

Solidity struct definition parsing support

token

ABI param and parsing for it.

Structs

Abi

API building calls to contracts ABI.

AbiError

Contract error specification.

AbiParser

A parser that turns a “human readable abi” into a Abi

Constructor

Contract constructor specification.

Contract

API building calls to contracts ABI.

Event

Contract event.

EventParam

Event param specification.

Events

Contract events iterator.

Function

Contract function specification.

Functions

Contract functions iterator.

InvalidOutputType

Log

Decoded log.

LogParam

Decoded log param.

Param

Function param.

RawLog

Ethereum log.

RawTopicFilter

Raw topic filter.

TopicFilter

Topic filter.

TupleParam

Tuple params specification

Enums

AbiError

ABI codec related errors

Error

Ethabi errors

ParamType

Function and event param types.

ParseError

Error that can occur during human readable parsing

StateMutability

Whether a function modifies or reads blockchain state

Token

Ethereum ABI params.

Topic

Acceptable topic possibilities.

Traits

AbiArrayType

Additional trait for types that can appear in arrays

AbiDecode

Trait for ABI decoding

AbiEncode

Trait for ABI encoding

AbiType

A trait for types that can be represented in the ethereum ABI.

Detokenize

Output type possible to deserialize from Contract ABI

EventExt

Extension trait for ethabi::Event.

FunctionExt

Extension trait for ethabi::Function.

FunctionOutputDecoder

Contract functions generated by ethabi-derive

LogFilter

Common filtering functions that are available for any event.

ParseLog

trait common to things (events) that have an associated Log type that can be parsed from a RawLog

Tokenizable

Simplified output type for single value.

TokenizableItem

Marker trait for Tokenizable types that are can tokenized to and from a Token::Array and Token:FixedArray.

Tokenize

Tokens conversion trait

Functions

decode

Decodes ABI compliant vector of bytes into vector of tokens described by types param.

encode

Encodes vector of tokens into ABI compliant vector of bytes.

long_signature

Returns the full Keccak-256 hash of the signature of the given params

parse_abi

Parses a “human readable abi” string vector

parse_abi_str

Parses a “human readable abi” string

short_signature

Returns the first four bytes of the Keccak-256 hash of the signature of the given params

Type Definitions

Address

ABI address.

Bytes

ABI bytes.

FixedBytes

ABI fixed bytes.

Hash

Commonly used FixedBytes of size 32

Int

ABI signed integer.

Result

Ethabi result type

Uint

ABI unsigned integer.

Word

ABI word.