Struct ethers_core::abi::Contract

pub struct Contract {
    pub constructor: Option<Constructor>,
    pub functions: BTreeMap<String, Vec<Function, Global>, Global>,
    pub events: BTreeMap<String, Vec<Event, Global>, Global>,
    pub errors: BTreeMap<String, Vec<Error, Global>, Global>,
    pub receive: bool,
    pub fallback: bool,
}

API building calls to contracts ABI.

Fields
constructor: Option<Constructor>

Contract constructor.

functions: BTreeMap<String, Vec<Function, Global>, Global>

Contract functions.

events: BTreeMap<String, Vec<Event, Global>, Global>

Contract events, maps signature to event.

errors: BTreeMap<String, Vec<Error, Global>, Global>

Contract errors, maps signature to error.

receive: bool

Contract has receive function.

fallback: bool

Contract has fallback function.

Implementations

impl Contract

pub fn load<T>(reader: T) -> Result<Contract, Error>where T: Read,

Loads contract from json.

pub fn constructor(&self) -> Option<&Constructor>

Creates constructor call builder.

pub fn function(&self, name: &str) -> Result<&Function, Error>

Get the function named name, the first if there are overloaded versions of the same function.

pub fn event(&self, name: &str) -> Result<&Event, Error>

Get the contract event named name, the first if there are multiple.

pub fn error(&self, name: &str) -> Result<&Error, Error>

Get the contract error named name, the first if there are multiple.

pub fn events_by_name(&self, name: &str) -> Result<&Vec<Event, Global>, Error>

Get all contract events named name.

pub fn functions_by_name( &self, name: &str ) -> Result<&Vec<Function, Global>, Error>

Get all functions named name.

pub fn errors_by_name(&self, name: &str) -> Result<&Vec<Error, Global>, Error>

Get all errors named name.

pub fn functions(&self) -> Functions<'_>

Iterate over all functions of the contract in arbitrary order.

pub fn events(&self) -> Events<'_>

Iterate over all events of the contract in arbitrary order.

pub fn errors(&self) -> AbiErrors<'_>

Iterate over all errors of the contract in arbitrary order.