Module ethers_core::types

Re-exports

pub use transaction::eip1559::Eip1559TransactionRequest;

pub use transaction::eip2930::Eip2930TransactionRequest;

pub use transaction::request::TransactionRequest;

pub use transaction::response::Transaction;

pub use transaction::response::TransactionReceipt;

Modules

transaction

Structs

AccountDiff

Serde-friendly AccountDiff shadow.

Block

The block type returned from RPC calls. This is generic over a TX type which will be either the hash or the

BlockTrace

Ad-Hoc trace API type

Bloom

Bloom hash type with 256 bytes (2048 bits) size.

Bytes

Wrapper type around Bytes to deserialize/serialize “0x” prefixed ethereum hex strings

Call

Call response

CallResult

Call Result

ChangedType

Aux type for Diff::Changed.

Create

Create response

CreateResult

Create Result

EIP1186ProofResponse

FeeHistory

Filter

Filter for

H160

Fixed-size uninterpreted hash type with 20 bytes (160 bits) size.

H256

Fixed-size uninterpreted hash type with 32 bytes (256 bits) size.

H512

Fixed-size uninterpreted hash type with 64 bytes (512 bits) size.

I256

Little-endian 256-bit signed integer.

Log

A log produced by a transaction.

MemoryDiff

A diff of some chunk of memory.

Reward

Reward action

Signature

An ECDSA signature

StateDiff

Serde-friendly StateDiff shadow.

StorageDiff

A diff of some storage value.

StorageProof

Suicide

Suicide

Trace

Trace-Filtering API trace type

TraceFilter

Trace filter

TransactionTrace

Trace

TxHash

A transaction Hash

TxpoolContent

Transaction Pool Content

TxpoolInspect

Transaction Pool Inspect

TxpoolInspectSummary

Transaction summary as found in the Txpool Inspection property.

TxpoolStatus

Transaction Pool Status

U64

Little-endian large integer type

U128

Little-endian large integer type

U256

Little-endian large integer type

VMExecutedOperation

A record of an executed VM operation.

VMOperation

A record of the execution of a single VM operation.

VMTrace

A record of a full VM trace for a CALL/CREATE.

Enums

Action

Action

ActionType

An external action type.

AddressOrBytes

A type that can either be an Address or Bytes.

BlockId

A Block Hash or Block Number

BlockNumber

A block Number (or tag - “latest”, “earliest”, “pending”)

CallType

Call type.

Chain

Diff

Serde-friendly Diff shadow.

FilterBlockOption

NameOrAddress

ENS name or Ethereum Address. Not RLP encoded/serialized if it’s a name

PathOrString

A type that can either be a Path or a String

RecoveryMessage

Recovery message data.

Res

Response

RewardType

Reward type.

Sign

Enum to represent the sign of a 256-bit signed integer.

SignatureError

An error involving a signature.

TraceType

Description of the type of trace to make

ValueOrArray

Union type for representing a single value or a vector of values inside a filter

Type Definitions

Address

Selector