Enum ethers_core::types::TraceError

#[repr(u8)]
pub enum TraceError {

    OutOfGas,
    BadJumpDestination,
    BadInstruction,
    StackUnderflow,
    OutOfStack,
    Internal,
    BuiltIn,
    MutableCallInStaticContext,
    Wasm,
    OutOfBounds,
    Reverted,
    SubStackUnderflow,
    OutOfSubStack,
    InvalidSubEntry,
    InvalidCode,

}

Trace evm errors.

Variants

OutOfGas

OutOfGas is returned when transaction execution runs out of gas.

BadJumpDestination

BadJumpDestination is returned when execution tried to move to position that wasn’t marked with JUMPDEST instruction

BadInstruction

BadInstructions is returned when given instruction is not supported

StackUnderflow

StackUnderflow when there is not enough stack elements to execute instruction

OutOfStack

When execution would exceed defined Stack Limit

Internal

Returned on evm internal error. Should never be ignored during development. Likely to cause consensus issues.

BuiltIn

When builtin contract failed on input data

MutableCallInStaticContext

When execution tries to modify the state in static context

Wasm

Wasm error

OutOfBounds

Contract tried to access past the return data buffer.

Reverted

Execution has been reverted with REVERT instruction.

SubStackUnderflow

When there is not enough subroutine stack elements to return from

OutOfSubStack

When execution would exceed defined subroutine Stack Limit

InvalidSubEntry

When the code walks into a subroutine, that is not allowed

InvalidCode

When invalid code was attempted to deploy

Implementations

impl TraceError

pub fn from_u8(val: u8) -> Result<Self, u8>

Converts the given byte into the corresponding error variant.