Enum ethers_core::types::transaction::eip2718::TypedTransaction

pub enum TypedTransaction {
    Legacy(TransactionRequest),
    Eip2930(Eip2930TransactionRequest),
    Eip1559(Eip1559TransactionRequest),
}

The TypedTransaction enum represents all Ethereum transaction types.

Its variants correspond to specific allowed transactions:

Legacy (pre-EIP2718) TransactionRequest

EIP2930 (state access lists) Eip2930TransactionRequest

EIP1559 Eip1559TransactionRequest

To support Kovan and other non-London-compatbile networks, please enable the legacy crate feature. This will disable the type flag in the serialized transaction, and cause contract calls and other common actions to default to using the legacy transaction type.

Variants

Legacy(TransactionRequest)

Eip2930(Eip2930TransactionRequest)

Eip1559(Eip1559TransactionRequest)

Implementations

impl TypedTransaction

pub fn from(&self) -> Option<&Address>

pub fn set_from(&mut self, from: Address) -> &mut Self

pub fn to(&self) -> Option<&NameOrAddress>

pub fn to_addr(&self) -> Option<&Address>

pub fn set_to<T: Into<NameOrAddress>>(&mut self, to: T) -> &mut Self

pub fn nonce(&self) -> Option<&U256>

pub fn set_nonce<T: Into<U256>>(&mut self, nonce: T) -> &mut Self

pub fn value(&self) -> Option<&U256>

pub fn set_value<T: Into<U256>>(&mut self, value: T) -> &mut Self

pub fn gas(&self) -> Option<&U256>

pub fn gas_mut(&mut self) -> &mut Option<U256>

pub fn set_gas<T: Into<U256>>(&mut self, gas: T) -> &mut Self

pub fn gas_price(&self) -> Option<U256>

pub fn set_gas_price<T: Into<U256>>(&mut self, gas_price: T) -> &mut Self

pub fn chain_id(&self) -> Option<U64>

pub fn set_chain_id<T: Into<U64>>(&mut self, chain_id: T) -> &mut Self

pub fn data(&self) -> Option<&Bytes>

pub fn access_list(&self) -> Option<&AccessList>

pub fn set_access_list(&mut self, access_list: AccessList) -> &mut Self

pub fn set_data(&mut self, data: Bytes) -> &mut Self

pub fn rlp_signed(&self, signature: &Signature) -> Bytes

pub fn rlp(&self) -> Bytes

pub fn sighash(&self) -> H256

Hashes the transaction’s data. Does not double-RLP encode

pub fn max_cost(&self) -> Option<U256>

Max cost of the transaction

pub fn hash(&self, signature: &Signature) -> H256

Hashes the transaction’s data with the included signature.

pub fn decode_signed( rlp: &Rlp<'_> ) -> Result<(Self, Signature), TypedTransactionError>

Decodes a signed TypedTransaction from a rlp encoded byte stream

impl TypedTransaction

pub fn as_legacy_ref(&self) -> Option<&TransactionRequest>

pub fn as_eip2930_ref(&self) -> Option<&Eip2930TransactionRequest>

pub fn as_eip1559_ref(&self) -> Option<&Eip1559TransactionRequest>

pub fn as_legacy_mut(&mut self) -> Option<&mut TransactionRequest>

pub fn as_eip2930_mut(&mut self) -> Option<&mut Eip2930TransactionRequest>

pub fn as_eip1559_mut(&mut self) -> Option<&mut Eip1559TransactionRequest>